Rename utilsnapshot -> utiltest
